Move code to src; remove Heroku configs
